refactor(server/client): storing vehicle images locally
The images/screenshots that are taken from vehicles are stored locally inside the es_extended not discord anymore. While using discord as storage, images start to disapper after couple of days, leaving us with no choice to regenerate vehicle data to be able to access their images. But won't be the case with this new refactored mathod, storing images localy.